tests/wpt/tests/css/css-backgrounds/parsing/background-size-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Backgrounds and Borders: getComputedStyle().backgroundSize</title>
<link rel="help" href="https://drafts.csswg.org/css-backgrounds/#background-size">
<meta name="assert" content="background-size computed value is a list, each item a pair of sizes (one per axis) each represented as either a keyword or a computed length-percentage value.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<style>
  #target {
    font-size: 40px;
  }
</style>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("background-size", "1px", "1px auto");
test_computed_value("background-size", "1px auto", "1px auto");
test_computed_value("background-size", "2% 3%");
test_computed_value("background-size", "auto");
test_computed_value("background-size", "auto auto", "auto");
test_computed_value("background-size", "auto 4%");
test_computed_value("background-size", "contain");
test_computed_value("background-size", "cover");
test_computed_value("background-size", "calc(10px + 0.5em) calc(10px - 0.5em)", "30px 0px");
test_computed_value("background-size", "calc(10px - 0.5em) calc(10px + 0.5em)", "0px 30px");

// See background-computed.html for a test with multiple background images.
test_computed_value("background-size", "auto 1px, 2% 3%, contain", "auto 1px");
</script>
</body>
</html>
